First working version
Based on the Python 3.8 grammar specification, this version of PseuToPy
allows to recreate some of Python's specific behaviors in terms of
assignments and combinations of operations. It also allows to use the
Python syntax to create instructions that we were not able to use
beforehand such as: set, dictionaries, lists, classes.

The current implementation allows for variable declarations and
assignements. Prints and inputs were also implemented. We now need to
implement funtion and method calls.